Sync coreth PR #957: move log to separate package (#1614)
