+++ latest-coreth-commit.txt
@@ -1,3 +1,3 @@
 # This is the latest commit of coreth that is synced with the subnet-evm
 
-3cf977c219ff46951c6b0a2227d69262a0a3038a
+8222fcf9d9fe0630538318c6e90235d8e78ecccd
